cleanups and python3 interop
Remove comments about what was changed where; these belong in the CONTRIB file.

Remove use of deprecated htmllib. Unfortunately the output is a little different
after this change. How different is yet to be investigated...

Load print from future and use the python version to determine which html parser
library to use. This makes the script work with both python 2 and 3.

Reformat to be pep8 compliant.

Remove comparison with old algorithm, as no current user will need that.